# Make the performance reporting update frequency customizable