--- a/core/script_language.h
+++ b/core/script_language.h
@@ -352,6 +352,8 @@ class ScriptDebugger {
 public:
 	typedef void (*RequestSceneTreeMessageFunc)(void *);
 
+	int update_frequency;
+
 	struct LiveEditFuncs {
 
 		void *udata;