--- a/main/main.h
+++ b/main/main.h
@@ -44,7 +44,7 @@ class Main {
 	static void print_help(const char *p_binary);
 	static uint64_t last_ticks;
 	static uint64_t target_ticks;
-	static uint32_t frames;
+	static Array frame_times;
 	static uint32_t frame;
 	static bool force_redraw_requested;