The default update frequency has been changed from 1000ms to 250ms. This makes FPS counter and monitor updates smoother and more reactive. I couldn't notice any significant performance decreases by lowering the default reporting rate to 250ms, even on a nearly empty project which ran at over 4,000 FPS on my machine.

I had to redesign the FPS counting algorithm after [this one](http://www.growingwiththeweb.com/2017/12/fast-simple-js-fps-counter.html) so that FPS counting no longer relies on an update frequency of 1000ms.

This implements https://github.com/godotengine/godot/issues/2794.