@akien-mga 2018-05-28
Would have liked @reduz to review this one before merging, because `Main::iteration()` is very critical for performance.