@reduz 2018-05-28
I reverted this, as this is not the right approach to solve the problem, and all it may cause is game stalls (which are happening already with the remote scene tree).

We learned from the remote scene tree than a different way to send this information over from the debugger is needed, probably from a thread. This needs more research before any of this should be touched.